Update and rename dist_percent_of_tot_const.cal to python_scripts/dist_percent_of_tot_const.py
This is an **.cal** file, saved from ArcMap's Field Calculator. It's only usable within that context using the Load function. Only the code within the "Pre-Logic Script" codeblock is to be copied into ArcMap. The below codeblock shows the expected output as to how ArcMap interprets the python code (note ArcMap also allows for VisualBasic code). 

Works in ArcMap 10.6.